requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coverage==7.2.1
cryptography==39.0.2
defusedxml==0.7.1
Django==2.1.5
django-allauth==0.52.0
django-cleanup==7.0.0
django-registration-redux==2.2
idna==3.4
oauthlib==3.2.2
Pillow==5.4.1
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
urllib3==1.26.14
